requirements.txt
requests
pyjwt
cryptography
python-dotenv